Optmize Code
